pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prjcap"
version = "0.1.0"
description = "Console task tracker for project-scoped AI agents"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "prjcap" }]

# Keep runtime deps minimal: Whisper/torch are expected to be available via your `act_env_general`.
dependencies = []

[project.optional-dependencies]
voice = ["openai-whisper", "torch"]

[project.scripts]
prjcap = "prjcap.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["prjcap"]

